refactor: PR レビュー指摘を反映
- GameProvider のマウント時に SSR と同じ GameDocument を即時 fetch
  していた問題を解消。useLazyQuery + setInterval に切り替え、初回も
  60 秒経過後に fetch する。
- useMyPlayer は useQuery の data を直接返すように整理し、useEffect
  は myPlayerAtom への書き込み専用にした（read-only 経路用）。
- usePollingPeriod は useGameValue() で現在の game を読むよう変更。
  ポーリングで status が変わった場合に追従する。
- messages-query.ts の re-export を撤去し、MessagesQueryProvider /
  useInitialMessagesQuery は contexts/messages-query-context から
  直接 import する形に統一。
