+++ frontend/src/components/pages/games/contexts/game-context.tsx
@@ -4,8 +4,14 @@ import {
   GameQuery,
   GameQueryVariables
 } from '@/lib/generated/graphql'
-import { useQuery } from '@apollo/client'
-import { ReactNode, createContext, useContext } from 'react'
+import { useLazyQuery } from '@apollo/client'
+import {
+  ReactNode,
+  createContext,
+  useContext,
+  useEffect,
+  useState
+} from 'react'
 
 const GameContext = createContext<Game | null>(null)
 
@@ -15,14 +21,26 @@ type Props = {
 }
 
 // 1分に1回 game を再取得し、参加者・期間・ステータスなどの変更を反映する。
+// useQuery を使うとマウント直後に SSR と同じ GameDocument を即時で再発行してしまうため、
+// useLazyQuery + setInterval で初回 fetch も 60s 後にずらす。
 const GAME_POLL_INTERVAL_MS = 60_000
 
 export const GameProvider = ({ game: initialGame, children }: Props) => {
-  const { data } = useQuery<GameQuery, GameQueryVariables>(GameDocument, {
-    variables: { id: initialGame.id },
-    pollInterval: GAME_POLL_INTERVAL_MS
-  })
-  const game = (data?.game as Game | undefined) ?? initialGame
+  const [game, setGame] = useState<Game>(initialGame)
+  const [refetchGame] = useLazyQuery<GameQuery, GameQueryVariables>(
+    GameDocument,
+    { variables: { id: initialGame.id } }
+  )
+
+  useEffect(() => {
+    const tick = async () => {
+      const { data } = await refetchGame()
+      if (data?.game) setGame(data.game as Game)
+    }
+    const timer = setInterval(tick, GAME_POLL_INTERVAL_MS)
+    return () => clearInterval(timer)
+  }, [refetchGame])
+
   return <GameContext.Provider value={game}>{children}</GameContext.Provider>
 }
